backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
tenacity==8.5.0
pydantic==2.9.2
chromadb==0.5.20
mem0ai==0.1.40
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
markdown==3.7
weasyprint==62.3